// src/interfaces/interfaces.ts
export interface Idata {
  num: string;
  name: string;
  count: string;
  year: string;
  shape: string;
  color: string;
  size: string;
  favorite: boolean;
}
export interface ISortData{
  children: {innerHTML: string}[];
  parentNode: NodeListOf<Element>;
}
export interface Iobj {
  num: string;
  name: string;
  count: string;
  year: string;
  shape: string;
  color: string;
  size: string;
  favorite: boolean;
}